[MRG] Automatic merge, trunk 17224 to 19763 ("Connection reset by peer" error mid-way)
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/branches/experimental@19765 b456876b-0849-0410-b77d-98878d47e9d5

+++ ui-revamp/doc/devtools/tikirelease.sh
@@ -1,72 +1,19 @@
-#!/bin/sh
+#!/bin/sh -x
 # $Id$
 #
-# ===========================================================
-# IMPORTANT NOTE : This script should not be called manually.
-#   It is called by doc/devtools/release.php
-# ===========================================================
+# This script does an export of a given subversion tree and creates a set of
+# release packages that can be uploaded to SourceForge and installed by end 
+# users.
+
+# ==========================================================================
+# IMPORTANT NOTE : This script must NOT be called directly !!!
+#   It is used by the main release script (doc/devtools/release.php)
+#   To get the Tiki release HOWTO, try: php doc/devtools/release.php --howto
+# ==========================================================================
 #
-# HOWTO release Tikiwiki ?
-# --------------------------
-# 
-# pre/
-#    - run doc/devtools/release_changelog.php to update changelog.txt and check the diff with the last version
-#    - run doc/devtools/release_copyright.php to update copyright.txt and check the diff with the last version
-#    - update README
-#    - Test the whole Installer and check if everything is OK
-#    - run doc/devtools/securitycheck.php and check each "potentially unsafe" file.
-#    - run doc/devtools/diffsql.sh to make sure tiki.sql and upgrade script from 
-#        previous version give the same db structure (but not necessarily the same data).
-#        The upgrade script is designed to be ran again & again. Specifically, we don't 
-#        want new permissions or modules to appear at upgrade. If there is a chance that 
-#        someone chose to delete something, it should not re-appear at each upgrade. 
-#    - cd db/convertscripts and run convertsqls.sh
-#    - in lib/setup/twversion.class.php
-#      - increment the version number in the constructor
-#      - update list of valid releases in getVersions()
-#    - commit your changes
-#
-# 1/ Create and test pre-release packages by executing the script with the release
-#    version as argument, using the format major.minor.sub 
-#    php doc/devtools/release.php 2.0 preRC4
-#
-# 2/ Test the produced tarball and share the testing : you need at least 3 install 
-#    from 3 different people
-# 
-# 3/ After testing, tag the release, build the release tarballs
-#    php doc/devtools/release.php 2.0 RC4
-#    
-# 4/ Test the produced tarball and share the testing : you need at least 3 install 
-#    from 3 different people
-# 
-# 5/ When the tarball is tested, follow the steps to upload on SourceForge:
-#    http://tinyurl.com/59uubv
-#    
-# 6/ Warn people that do .rpm and ebuilds that the archive is avalaible so they can
-#    complete the packaging process with new files. If you don't know who does that,
-#    warn everybody.
-#
-# 7/ unless in step 8/ you warned everybody you have now to announce the good news
-#    on devel mailing-list and ask marc to launch the announce-speading process 
-#    (Freshmeat, SourceForge and tikiwiki.org (manually for now).
-#
-# post/ After release, update templates/tiki-install.tpl and 
-#       templates/tiki-top_bar.tpl (including templates/styles/*/tiki-top_bar.tpl) 
-#       to next version number with CVS   ex.: 1.9.2 (CVS)  . This helps later on to 
-#       know exactly which files were included or not in a release.
-#
-#		- Also, update appropriate tw.o/*.version file with new release version
-#
-#
-# All that process has to be relayed on live irc channel : 
-# irc://irc.freenode.net/#tikiwiki
-#
-# ############################################################
-# start of configuration
-# change here what you need to fit your environment
 
 SVNROOT="https://tikiwiki.svn.sourceforge.net/svnroot/tikiwiki"
-WORKDIR="/home/$USER/tikipack"
+WORKDIR="$HOME/tikipack"
 MODULE="tikiwiki"
 
 # end of configuration
@@ -79,40 +26,53 @@ echo "Usage: tikirelease.sh <release-version> <svn-relative-path>"
 	exit 0
 fi
 
-OLDIR=`pwd`
+OLDIR="`pwd`"
 VER=$1
 RELTAG=$2
 
 # ############################################################
 
-mkdir -p $WORKDIR
-cd $WORKDIR
+if [ ! -d $WORKDIR ]; then
+    mkdir -p $WORKDIR || die "Can't make $WORKDIR - $!"
+fi
+
+cd $WORKDIR || die "Can't get into $WORKDIR - $!"
+echo "Working in $WORKDIR"
 
 if [ -d $VER ]; then
-	rm -rf $VER
+    echo "Deleting old $VER"
+    rm -rf $VER
 fi
 mkdir $VER
 cd $VER
+
+echo "Exporting $SVNROOT/$RELTAG $MODULE-$VER"
 svn export $SVNROOT/$RELTAG $MODULE-$VER
+
+echo "Cleaning up"
 find $MODULE-$VER -name .cvsignore -type f -exec rm -f {} \;
 find $MODULE-$VER -name .svnignore -type f -exec rm -f {} \;
-find $MODULE-$VER -type d -exec chmod 775 {} \;
-find $MODULE-$VER -type f -exec chmod 664 {} \;
 
-# some more cleanup
 rm -rf $MODULE-$VER/tests
 rm -rf $MODULE-$VER/db/convertscripts
 rm -rf $MODULE-$VER/doc/devtools
 
+echo "Setting permissions"
+find $MODULE-$VER -type d -exec chmod 775 {} \;
+find $MODULE-$VER -type f -exec chmod 664 {} \;
+
+echo "Creating tarballs"
 tar -czf $MODULE-$VER.tar.gz $MODULE-$VER
 tar -cjf $MODULE-$VER.tar.bz2 $MODULE-$VER
 zip -r $MODULE-$VER.zip $MODULE-$VER
 
+ls $WORKDIR/$VER
+
 echo ""
-echo "To upload the archives, copy-paste and exectue the following line at will (depending on SF's mood):"
-echo "cd $WORKDIR/$VER; lftp -u anonymous,release@tikiwiki.org -e 'cd incoming;put $MODULE-$VER.tar.gz;put $MODULE-$VER.tar.bz2;put $MODULE-$VER.zip;quit;' upload.sf.net"
+echo "To upload the 'tarballs', copy-paste and execute the following line (and change '\$SF_LOGIN' by your SF.net login):"
+echo "cd $WORKDIR/$VER; scp $MODULE-$VER.tar.gz $MODULE-$VER.tar.bz2 $MODULE-$VER.zip \$SF_LOGIN@frs.sourceforge.net:uploads"
 echo ""
 
-cd $OLDIR
+cd "$OLDIR"
 
 echo "Done."
